Continue fixing compilation errors: icons, smart casts, method signatures
- Fix remaining icon references in PodcastPlayerScreen and MusicPlayerScreen
- Add missing imports for MediaItem and clickable
- Fix smart cast issues by using safe let blocks
- Remove deprecated onValueChangeStarted parameters
- Update service method calls to match signatures
- Fix ListItem syntax with proper clickable modifier

+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/audiobook/AudiobookPlayerScreen.kt
@@ -185,7 +185,7 @@ fun AudiobookPlayerScreen(
             SleepTimerDialog(
                 currentTimer = audiobookState.sleepTimerEndTime,
                 onTimerSet = { minutes ->
-                    viewModel.setSleepTimer(minutes)
+                    viewModel.setSleepTimer(minutes.toInt())
                     showSleepTimer = false
                 },
                 onDismiss = { showSleepTimer = false }
